Move text status emitter into a separate module
